Expand Tasks to Calls in order to allow functions with outputs
